starwars/requirements.txt
aiohttp==2.3.1
asgi-redis==1.4.3
asgiref==1.1.2
asn1crypto==0.24.0
astroid==1.6.0
async-timeout==2.0.0
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
autopep8==1.3.3
cffi==1.11.4
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.4
daphne==1.4.2
Django==1.11.6
django-filter==1.1.0
django-filters==0.2.1
django-graph-auth==0.3.1
django-jwt-auth==0.0.2
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
graphene==2.0.1
graphene-django==2.0.0
graphql-core==2.0
graphql-relay==0.4.5
graphql-ws==0.2.0
h2==3.0.1
hpack==3.0.0
hyperframe==5.1.0
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
iso8601==0.1.12
isort==4.2.15
lazy-object-proxy==1.3.1
Markdown==2.6.9
mccabe==0.6.1
msgpack-python==0.5.1
multidict==4.0.0
priority==1.3.0
promise==2.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pycparser==2.18
PyJWT==1.5.3
pylint==1.8.1
pyOpenSSL==17.5.0
pytz==2017.2
redis==2.10.6
Rx==1.6.0
service-identity==17.0.0
singledispatch==3.4.0.3
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
typing==3.6.2
wheel==0.24.0
wrapt==1.10.11
yarl==1.0.0
zope.interface==4.4.3
